FavoriteLogAction.java
